add context for auth
