+++ Context/Data/Entities/Role.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Auth.Models;
-
+namespace Auth.Data.Entities;
 
 [Table("Role", Schema = "dbo")]
 public class Role
